feat: create module for auth-related endpoints
