+++ backend/views/auth.py
@@ -1,8 +1,8 @@
 from flask import make_response
 from flask_restful import Api, Resource, request
 from marshmallow import ValidationError
-from repos.users import UsersRepo
-from schemas.users import LoginUserSchema, UserSchema
+from repos import AuthRepo, UsersRepo
+from schemas import GithubAuthSchema, LoginUserSchema, UserSchema
 from services.decorators import validate
 from views.blueprints import main_bp
 
@@ -37,6 +37,19 @@ class LoginUser(Resource):
             return {"error": "Validation failed", "messages": e.messages}, 400
 
 
+class GenerateAccessToken(Resource):
+    def post(self):
+        try:
+            auth_serializer = GithubAuthSchema()
+            auth_data = auth_serializer.load(request.get_json())
+            res = AuthRepo.get_access_token(auth_data["code"])
+            if res["error"]:
+                return res, 400
+            return auth_serializer.dump(res), 200
+        except ValidationError as e:
+            return {"error": "Validation failed", "messages": e.messages}, 400
+
+
 # can delete later
 class TestAuthGuard(Resource):
     @validate
@@ -47,3 +60,4 @@ class TestAuthGuard(Resource):
 main_api.add_resource(RegisterUser, "/register")
 main_api.add_resource(LoginUser, "/login")
 main_api.add_resource(TestAuthGuard, "/test")
+main_api.add_resource(GenerateAccessToken, "/generate-access-token")
